New show command. Old show is now open.
Since the layout of ls was altered, there was need for some command to show a
molecule with all its atoms. That is show now. The command to open a x-fileref
atom in the desktop default application is open. For this all to work properly
the API of IMoleculeViewer had to be adjusted to differentiate between varying
verbosities of output.